19458: add sentenceBERT eval script
